API-ML/requirements.txt
Flask
gunicorn
tensorflow
keras
joblib
pillow
firebase
firebase_admin
google-cloud-core
google-cloud-firestore
google-cloud-storage